+++ source/isaaclab/isaaclab/scene/interactive_scene.py
@@ -12,6 +12,7 @@ import omni.log
 import omni.usd
 from isaacsim.core.cloner import GridCloner
 from isaacsim.core.prims import XFormPrim
+from isaacsim.core.utils.stage import get_current_stage, get_current_stage_id
 from pxr import PhysxSchema
 
 import isaaclab.sim as sim_utils
@@ -25,8 +26,11 @@ from isaaclab.assets import (
     RigidObjectCfg,
     RigidObjectCollection,
     RigidObjectCollectionCfg,
+    SurfaceGripper,
+    SurfaceGripperCfg,
 )
 from isaaclab.sensors import ContactSensorCfg, FrameTransformerCfg, SensorBase, SensorBaseCfg
+from isaaclab.sim import SimulationContext
 from isaaclab.terrains import TerrainImporter, TerrainImporterCfg
 
 from .interactive_scene_cfg import InteractiveSceneCfg
@@ -118,13 +122,16 @@ class InteractiveScene:
         self._rigid_objects = dict()
         self._rigid_object_collections = dict()
         self._sensors = dict()
+        self._surface_grippers = dict()
         self._extras = dict()
-        # obtain the current stage
-        self.stage = omni.usd.get_context().get_stage()
+        # get stage handle
+        self.sim = SimulationContext.instance()
+        self.stage = get_current_stage()
+        self.stage_id = get_current_stage_id()
         # physics scene path
         self._physics_scene_path = None
         # prepare cloner for environment replication
-        self.cloner = GridCloner(spacing=self.cfg.env_spacing)
+        self.cloner = GridCloner(spacing=self.cfg.env_spacing, stage=self.stage)
         self.cloner.define_base_env(self.env_ns)
         self.env_prim_paths = self.cloner.generate_paths(f"{self.env_ns}/env", self.cfg.num_envs)
         # create source prim
@@ -340,6 +347,11 @@ class InteractiveScene:
         return self._sensors
 
     @property
+    def surface_grippers(self) -> dict[str, SurfaceGripper]:
+        """A dictionary of the surface grippers in the scene."""
+        return self._surface_grippers
+
+    @property
     def extras(self) -> dict[str, XFormPrim]:
         """A dictionary of miscellaneous simulation objects that neither inherit from assets nor sensors.
 
@@ -384,6 +396,8 @@ class InteractiveScene:
             deformable_object.reset(env_ids)
         for rigid_object in self._rigid_objects.values():
             rigid_object.reset(env_ids)
+        for surface_gripper in self._surface_grippers.values():
+            surface_gripper.reset(env_ids)
         for rigid_object_collection in self._rigid_object_collections.values():
             rigid_object_collection.reset(env_ids)
         # -- sensors
@@ -399,6 +413,8 @@ class InteractiveScene:
             deformable_object.write_data_to_sim()
         for rigid_object in self._rigid_objects.values():
             rigid_object.write_data_to_sim()
+        for surface_gripper in self._surface_grippers.values():
+            surface_gripper.write_data_to_sim()
         for rigid_object_collection in self._rigid_object_collections.values():
             rigid_object_collection.write_data_to_sim()
 
@@ -417,6 +433,8 @@ class InteractiveScene:
             rigid_object.update(dt)
         for rigid_object_collection in self._rigid_object_collections.values():
             rigid_object_collection.update(dt)
+        for surface_gripper in self._surface_grippers.values():
+            surface_gripper.update(dt)
         # -- sensors
         for sensor in self._sensors.values():
             sensor.update(dt, force_recompute=not self.cfg.lazy_sensor_update)
@@ -479,6 +497,10 @@ class InteractiveScene:
             root_velocity = asset_state["root_velocity"].clone()
             rigid_object.write_root_pose_to_sim(root_pose, env_ids=env_ids)
             rigid_object.write_root_velocity_to_sim(root_velocity, env_ids=env_ids)
+        # surface grippers
+        for asset_name, gripper in self._surface_grippers.items():
+            asset_state = state["gripper"][asset_name]
+            gripper.write_gripper_state_to_sim(asset_state, env_ids=env_ids)
 
         # write data to simulation to make sure initial state is set
         # this propagates the joint targets to the simulation
@@ -584,6 +606,7 @@ class InteractiveScene:
             self._rigid_objects,
             self._rigid_object_collections,
             self._sensors,
+            self._surface_grippers,
             self._extras,
         ]:
             all_keys += list(asset_family.keys())
@@ -610,6 +633,7 @@ class InteractiveScene:
             self._rigid_objects,
             self._rigid_object_collections,
             self._sensors,
+            self._surface_grippers,
             self._extras,
         ]:
             out = asset_family.get(key)
@@ -668,6 +692,8 @@ class InteractiveScene:
                     if hasattr(rigid_object_cfg, "collision_group") and rigid_object_cfg.collision_group == -1:
                         asset_paths = sim_utils.find_matching_prim_paths(rigid_object_cfg.prim_path)
                         self._global_prim_paths += asset_paths
+            elif isinstance(asset_cfg, SurfaceGripperCfg):
+                pass
             elif isinstance(asset_cfg, SensorBaseCfg):
                 # Update target frame path(s)' regex name space for FrameTransformer
                 if isinstance(asset_cfg, FrameTransformerCfg):